Migrate to javascript action (instead of composite) (#10)
Previous version of the action was of type "composite", with javascript code written directly into the action.yml file.
In this PR I've re-written the action to be of type "javascript" instead.

By having the code in actual *.js or *.ts files we get way better intellisense, and also have the possibility to introduce tests, linting, etc.

At the moment the action need to be manually bundled using npm run package script.